fix(pr): improvements
